separate plugin dir evaluation
To speed up ocm context creation with plugins, the evaluation
of a plugin dir is put into a separate cache now.
Context creation then just reuses known plugin dir setup
to create a context binding without calling all the plugins.
