+++ pkg/contexts/ocm/attrs/plugincacheattr/attr.go
@@ -9,6 +9,7 @@ import (
 	"github.com/open-component-model/ocm/pkg/contexts/ocm"
 	"github.com/open-component-model/ocm/pkg/contexts/ocm/attrs/plugindirattr"
 	"github.com/open-component-model/ocm/pkg/contexts/ocm/plugin/cache"
+	"github.com/open-component-model/ocm/pkg/contexts/ocm/plugin/plugins"
 )
 
 const (
@@ -17,15 +18,15 @@ const (
 
 ////////////////////////////////////////////////////////////////////////////////
 
-func Get(ctx ocm.Context) cache.Cache {
+func Get(ctx ocm.Context) plugins.Set {
 	path := plugindirattr.Get(ctx)
 
 	// avoid dead lock reading attribute during attribute creation
 	return ctx.GetAttributes().GetOrCreateAttribute(ATTR_KEY, func(ctx datacontext.Context) interface{} {
-		return cache.New(ctx.(ocm.Context), path)
-	}).(cache.Cache)
+		return plugins.New(ctx.(ocm.Context), path)
+	}).(plugins.Set)
 }
 
-func Set(ctx ocm.Context, cache cache.Cache) error {
+func Set(ctx ocm.Context, cache cache.PluginDir) error {
 	return ctx.GetAttributes().SetAttribute(ATTR_KEY, cache)
 }
