+++ pkg/contexts/ocm/plugin/const.go
@@ -1,0 +1,11 @@
+// SPDX-FileCopyrightText: 2022 SAP SE or an SAP affiliate company and Open Component Model contributors.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package plugin
+
+import (
+	"github.com/mandelsoft/logging"
+)
+
+var PKG = logging.Package()
